Refactor summary to different class
