+++ src/PlayerBanInfo.cs
@@ -1,7 +1,11 @@
-﻿namespace SharpSteamWebApi
+﻿using System;
+using System.Linq;
+using System.Xml.Linq;
+
+namespace SharpSteamWebApi
 {
     // Steam player ban info container.
-    public class PlayerBanInfo
+    public class PlayerBanInfo : SSWAObject
     {
         public bool HasVACBan { get; set; }
         public bool HasCommunityBan { get; set; }
@@ -37,6 +41,38 @@
         public bool HasDaysSinceLastBan()
         {
             return DaysSinceLastBan != -1;
+        }
+
+        // Queries player bans.
+        public static PlayerBanInfo Query(string apikey, long playerId)
+        {
+            string url = String.Format("http://api.steampowered.com/ISteamUser/GetPlayerBans/v1/?key={0}&steamids={1}&format=xml", apikey, playerId);
+            XDocument xml = GetXML(url);
+
+            if (xml == null)
+                return new PlayerBanInfo();
+
+            return Parse(xml.Descendants("player").ToArray()[0]);
+        }
+
+        // Parses player bans
+        private static PlayerBanInfo Parse(XElement xml)
+        {
+            if (xml == null)
+                return new PlayerBanInfo();
+
+            ElementParser parser = new ElementParser(xml);
+            PlayerBanInfo result = new PlayerBanInfo
+            {
+                HasVACBan = parser.GetAttributeBoolean("VACBanned"),
+                HasCommunityBan = parser.GetAttributeBoolean("CommunityBanned"),
+                HasTradeBan = parser.GetAttributeBoolean("EconomyBan"),
+                VACBanCount = parser.GetAttributeInteger("NumberOfVACBans"),
+                GameBanCount = parser.GetAttributeInteger("NumberOfGameBans"),
+                DaysSinceLastBan = parser.GetAttributeInteger("DaysSinceLastBan")
+            };
+
+            return result;
         }
     }
 }
