+++ src/SharpSteamWebApi.csproj
@@ -46,6 +46,7 @@
     <Compile Include="Game.cs" />
     <Compile Include="NewsArticle.cs" />
     <Compile Include="Player.cs" />
+    <Compile Include="PlayerSummary.cs" />
     <Compile Include="SSWAObject.cs" />
     <Compile Include="Statistic.cs" />
     <Compile Include="SteamWebApi.cs" />
